refactor(cursor): fold the whole cursor system into one component
CustomCursor now owns everything cursor-shaped: the inspectable hover
label, the cursor chat, and remote visitors' cursors, all fed by a
single mousemove listener (raw position for chat, edge-flipped for the
hover label) and the shared CursorLabel treatment. cursor-chat.tsx,
remote-cursors.tsx and cursor-label.tsx fold into custom-cursor/index
and canvas-layer mounts just <CustomCursor />. Remote cursor visuals
are unchanged. Supabase stays in the lazy chunk - the merged component
only reads the realtime store.
