+++ src/components/layout/canvas-layer.tsx
@@ -10,8 +10,6 @@ import {
   AppLoadingHandler,
   useAppLoadingStore
 } from "@/components/loading/app-loading-handler"
-import { CursorChat } from "@/components/realtime/cursor-chat"
-import { RemoteCursors } from "@/components/realtime/remote-cursors"
 import { useCanvasAvailability } from "@/hooks/use-canvas-availability"
 import { markCanvasBootStage } from "@/lib/canvas-boot"
 import { cn } from "@/utils/cn"
@@ -43,8 +41,6 @@ export const CanvasLayer = () => {
     <>
       <div className="pointer-events-none fixed top-0 z-50 h-screen w-full">
         <CustomCursor />
-        <RemoteCursors />
-        <CursorChat />
       </div>
 
       {/* Dead weight without a renderer: an invisible overlay and a 3D-only viewer. */}
